+++ feature-libs/user/account/components/my-account-v2-user/index.ts
@@ -4,5 +4,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './myaccount-v2-user.component';
-export * from './myaccount-v2-user.module';
+export * from './my-account-v2-user.component';
+export * from './my-account-v2-user.module';
