+++ feature-libs/user/account/components/my-account-v2-user/my-account-v2-user.component.html
@@ -1,4 +1,4 @@
-<div class="cx-myaccount-v2-user">
+<div class="cx-my-account-v2-user">
   <ng-container *ngIf="user$ | async as user">
     <div class="cx-name">{{ user.title }} {{ user.name }}</div>
     <a
