+++ feature-libs/user/account/components/my-account-v2-user/my-account-v2-user.component.ts
@@ -8,7 +8,7 @@ import { Component } from '@angular/core';
 import { LoginComponent } from '../login';
 
 @Component({
-  selector: 'cx-myaccount-v2-user',
-  templateUrl: './myaccount-v2-user.component.html',
+  selector: 'cx-my-account-v2-user',
+  templateUrl: './my-account-v2-user.component.html',
 })
 export class MyAccountV2UserComponent extends LoginComponent {}
