+++ feature-libs/user/account/components/my-account-v2-user/my-account-v2-user.module.ts
@@ -14,7 +14,7 @@ import {
   provideDefaultConfig,
 } from '@spartacus/core';
 import { RouterModule } from '@angular/router';
-import { MyAccountV2UserComponent } from './myaccount-v2-user.component';
+import { MyAccountV2UserComponent } from './my-account-v2-user.component';
 
 @NgModule({
   providers: [
