+++ feature-libs/user/account/components/public_api.ts
@@ -8,4 +8,4 @@ export * from './login-form/index';
 export * from './login-register/index';
 export * from './login/index';
 export * from './user-account-component.module';
-export * from './myaccount-v2-user/index';
+export * from './my-account-v2-user/index';
