+++ feature-libs/user/account/components/user-account-component.module.ts
@@ -8,7 +8,7 @@ import { NgModule } from '@angular/core';
 import { LoginFormModule } from './login-form/login-form.module';
 import { LoginRegisterModule } from './login-register/login-register.module';
 import { LoginModule } from './login/login.module';
-import { MyAccountV2UserModule } from './myaccount-v2-user';
+import { MyAccountV2UserModule } from './my-account-v2-user';
 
 @NgModule({
   imports: [
