+++ feature-libs/user/account/styles/_index.scss
@@ -1,3 +1,3 @@
 @import './login';
 @import './login-form';
-@import './myaccount-v2-user';
+@import './my-account-v2-user';
