+++ feature-libs/user/account/styles/_my-account-v2-user.scss
@@ -1,4 +1,4 @@
-.cx-myaccount-v2-user {
+.cx-my-account-v2-user {
   border: 1px solid var(--cx-color-light);
   width: 250px;
   padding: 20px 5px 5px 25px;
