+++ projects/storefrontlib/cms-components/myaccount/index.ts
@@ -12,6 +12,6 @@ export * from './my-interests/my-interests.module';
 export * from './notification-preference/index';
 export * from './payment-methods/payment-methods.component';
 export * from './payment-methods/payment-methods.module';
-export * from './myaccount-v2-navigation/myaccount-v2-navigation.component';
-export * from './myaccount-v2-navigation/myaccount-v2-navigation.module';
-export * from './myaccount-v2-navigation/index';
+export * from './my-account-v2-navigation/my-account-v2-navigation.component';
+export * from './my-account-v2-navigation/my-account-v2-navigation.module';
+export * from './my-account-v2-navigation/index';
