+++ projects/storefrontlib/cms-components/myaccount/my-account-v2-navigation/index.ts
@@ -4,5 +4,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './myaccount-v2-navigation.component';
-export * from './myaccount-v2-navigation.module';
+export * from './my-account-v2-navigation.component';
+export * from './my-account-v2-navigation.module';
