+++ projects/storefrontstyles/scss/components/content/_index.scss
@@ -9,4 +9,4 @@
 $content-components-allowlist: cx-banner, cx-link, cx-media, cx-carousel,
   cx-banner-carousel, cx-global-message, cx-navigation-ui, cx-footer-navigation,
   cx-category-navigation, cx-breadcrumb, cx-page-header, cx-scroll-to-top,
-  cx-video, cx-pdf, cx-myaccount-v2-navigation !default;
+  cx-video, cx-pdf, cx-my-account-v2-navigation !default;
