+++ projects/storefrontstyles/scss/components/content/navigation/_index.scss
@@ -4,4 +4,4 @@
 @import './category-navigation';
 @import './footer-navigation';
 @import './scroll-to-top';
-@import './myaccount-v2-navigation';
+@import './my-account-v2-navigation';
