+++ projects/storefrontstyles/scss/components/content/navigation/_my-account-v2-navigation.scss
@@ -1,4 +1,4 @@
-%cx-myaccount-v2-navigation {
+%cx-my-account-v2-navigation {
   background: var(--cx-color-inverse);
   font-family: sans-serif;
   padding: 5px 5px 5px 5px;
